Move ubjson implementation into internal header
This facilitates inlining into amspipe and also enables templating.
